+++ examples/Serial.cpp
@@ -19,7 +19,7 @@ public:
 	 * this is only a simple example. Messages should normally be
 	 * queued for later output when the application is less busy.
 	 */
-	void add_log_message(std::shared_ptr<uuid::log::Message> message) {
+	void operator<<(std::shared_ptr<uuid::log::Message> message) {
 		Serial.printf_P(PSTR("%s %c [%S] %s\r\n"),
 				uuid::log::format_timestamp_ms(message->uptime_ms, 3).c_str(),
 				uuid::log::format_level_char(message->level),